// Brorep.Ui/ClientApp/src/app/brorep-api.ts
export class WorkoutDto implements IWorkoutDto {
    workoutId?: string;
    name?: string | undefined;
    description?: string | undefined;
    imageUrl?: string | undefined;

    constructor(data?: IWorkoutDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.workoutId = data["workoutId"];
            this.name = data["name"];
            this.description = data["description"];
            this.imageUrl = data["imageUrl"];
        }
    }

    static fromJS(data: any): WorkoutDto {
        data = typeof data === 'object' ? data : {};
        let result = new WorkoutDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["workoutId"] = this.workoutId;
        data["name"] = this.name;
        data["description"] = this.description;
        data["imageUrl"] = this.imageUrl;
        return data; 
    }
}

export interface IWorkoutDto {
    workoutId?: string;
    name?: string | undefined;
    description?: string | undefined;
    imageUrl?: string | undefined;
}

export class SubmitWorkoutCommand implements ISubmitWorkoutCommand {
    username?: string | undefined;
    workoutId?: string;
    reps?: RepDto[] | undefined;
    videoUrl?: string | undefined;

    constructor(data?: ISubmitWorkoutCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.username = data["username"];
            this.workoutId = data["workoutId"];
            if (data["reps"] && data["reps"].constructor === Array) {
                this.reps = [] as any;
                for (let item of data["reps"])
                    this.reps!.push(RepDto.fromJS(item));
            }
            this.videoUrl = data["videoUrl"];
        }
    }

    static fromJS(data: any): SubmitWorkoutCommand {
        data = typeof data === 'object' ? data : {};
        let result = new SubmitWorkoutCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["username"] = this.username;
        data["workoutId"] = this.workoutId;
        if (this.reps && this.reps.constructor === Array) {
            data["reps"] = [];
            for (let item of this.reps)
                data["reps"].push(item.toJSON());
        }
        data["videoUrl"] = this.videoUrl;
        return data; 
    }
}

export interface ISubmitWorkoutCommand {
    username?: string | undefined;
    workoutId?: string;
    reps?: RepDto[] | undefined;
    videoUrl?: string | undefined;
}

export class RepDto implements IRepDto {
    startTime?: number;
    stopTime?: number;

    constructor(data?: IRepDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.startTime = data["startTime"];
            this.stopTime = data["stopTime"];
        }
    }

    static fromJS(data: any): RepDto {
        data = typeof data === 'object' ? data : {};
        let result = new RepDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["startTime"] = this.startTime;
        data["stopTime"] = this.stopTime;
        return data; 
    }
}

export interface IRepDto {
    startTime?: number;
    stopTime?: number;
}